+++ src/components/Footer/Footer.tsx
@@ -1,7 +1,7 @@
 import React, { Component } from 'react';
 import './Footer.css';
 
-export default class Footer extends Component {
+export class Footer extends Component {
     render() {
         return (
             <footer className="footer">
