+++ src/components/Navigation.css
@@ -11,6 +11,7 @@
     font-size: 1.2rem;
     font-weight: 500;
     color: #ffffff;
+    text-transform: uppercase;
 }
 
 .nav__btn {
