+++ src/components/data/productsData.ts
@@ -8,7 +8,8 @@ export const products: IProduct[] = [
         description:
             'Your perfect pack for everyday use and walks in the forest. Stash your laptop (up to 15 inches) in the padded sleeve, your everyday',
         category: "men's clothing",
-        image: 'https://fakestoreapi.com/img/81fPKd-2AYL._AC_SL1500_.jpg',
+        // image: 'https://fakestoreapi.com/img/81fPKd-2AYL._AC_SL1500_.jpg',
+        image: '../../assets/data-img/item_1.jpg',
         rating: {
             rate: 3.9,
             count: 120,
@@ -21,7 +22,8 @@ export const products: IProduct[] = [
         description:
             'Slim-fitting style, contrast raglan long sleeve, three-button henley placket, light weight & soft fabric for breathable and comfortable wearing. And Solid stitched shirts with round neck made for durability and a great fit for casual fashion wear and diehard baseball fans. The Henley style round neckline includes a three-button placket.',
         category: "men's clothing",
-        image: 'https://fakestoreapi.com/img/71-3HjGNDUL._AC_SY879._SX._UX._SY._UY_.jpg',
+        // image: 'https://fakestoreapi.com/img/71-3HjGNDUL._AC_SY879._SX._UX._SY._UY_.jpg',
+        image: '../../assets/data-img/item_2.jpg',
         rating: {
             rate: 4.1,
             count: 259,
@@ -34,7 +36,8 @@ export const products: IProduct[] = [
         description:
             'great outerwear jackets for Spring/Autumn/Winter, suitable for many occasions, such as working, hiking, camping, mountain/rock climbing, cycling, traveling or other outdoors. Good gift choice for you or your family member. A warm hearted love to Father, husband or son in this thanksgiving or Christmas Day.',
         category: "men's clothing",
-        image: 'https://fakestoreapi.com/img/71li-ujtlUL._AC_UX679_.jpg',
+        // image: 'https://fakestoreapi.com/img/71li-ujtlUL._AC_UX679_.jpg',
+        image: '../../assets/data-img/item_3.jpg',
         rating: {
             rate: 4.7,
             count: 500,
@@ -47,7 +50,8 @@ export const products: IProduct[] = [
         description:
             'The color could be slightly different between on the screen and in practice. / Please note that body builds vary by person, therefore, detailed size information should be reviewed below on the product description.',
         category: "men's clothing",
-        image: 'https://fakestoreapi.com/img/71YXzeOuslL._AC_UY879_.jpg',
+        // image: 'https://fakestoreapi.com/img/71YXzeOuslL._AC_UY879_.jpg',
+        image: '../../assets/data-img/item_4.jpg',
         rating: {
             rate: 2.1,
             count: 430,
@@ -60,7 +64,8 @@ export const products: IProduct[] = [
         description:
             "From our Legends Collection, the Naga was inspired by the mythical water dragon that protects the ocean's pearl. Wear facing inward to be bestowed with love and abundance, or outward for protection.",
         category: 'jewelery',
-        image: 'https://fakestoreapi.com/img/71pWzhdJNwL._AC_UL640_QL65_ML3_.jpg',
+        // image: 'https://fakestoreapi.com/img/71pWzhdJNwL._AC_UL640_QL65_ML3_.jpg',
+        image: '../../assets/data-img/item_5.jpg',
         rating: {
             rate: 4.6,
             count: 400,
@@ -73,7 +78,8 @@ export const products: IProduct[] = [
         description:
             'Satisfaction Guaranteed. Return or exchange any order within 30 days.Designed and sold by Hafeez Center in the United States. Satisfaction Guaranteed. Return or exchange any order within 30 days.',
         category: 'jewelery',
-        image: 'https://fakestoreapi.com/img/61sbMiUnoGL._AC_UL640_QL65_ML3_.jpg',
+        // image: 'https://fakestoreapi.com/img/61sbMiUnoGL._AC_UL640_QL65_ML3_.jpg',
+        image: '../../assets/data-img/item_6.jpg',
         rating: {
             rate: 3.9,
             count: 70,
@@ -86,7 +92,8 @@ export const products: IProduct[] = [
         description:
             "Classic Created Wedding Engagement Solitaire Diamond Promise Ring for Her. Gifts to spoil your love more for Engagement, Wedding, Anniversary, Valentine's Day...",
         category: 'jewelery',
-        image: 'https://fakestoreapi.com/img/71YAIFU48IL._AC_UL640_QL65_ML3_.jpg',
+        // image: 'https://fakestoreapi.com/img/71YAIFU48IL._AC_UL640_QL65_ML3_.jpg',
+        image: '../../assets/data-img/item_7.jpg',
         rating: {
             rate: 3,
             count: 400,
@@ -98,7 +105,8 @@ export const products: IProduct[] = [
         price: 10.99,
         description: 'Rose Gold Plated Double Flared Tunnel Plug Earrings. Made of 316L Stainless Steel',
         category: 'jewelery',
-        image: 'https://fakestoreapi.com/img/51UDEzMJVpL._AC_UL640_QL65_ML3_.jpg',
+        // image: 'https://fakestoreapi.com/img/51UDEzMJVpL._AC_UL640_QL65_ML3_.jpg',
+        image: '../../assets/data-img/item_8.jpg',
         rating: {
             rate: 1.9,
             count: 100,
@@ -111,7 +119,8 @@ export const products: IProduct[] = [
         description:
             'USB 3.0 and USB 2.0 Compatibility Fast data transfers Improve PC Performance High Capacity; Compatibility Formatted NTFS for Windows 10, Windows 8.1, Windows 7; Reformatting may be required for other operating systems; Compatibility may vary depending on user’s hardware configuration and operating system',
         category: 'electronics',
-        image: 'https://fakestoreapi.com/img/61IBBVJvSDL._AC_SY879_.jpg',
+        // image: 'https://fakestoreapi.com/img/61IBBVJvSDL._AC_SY879_.jpg',
+        image: '../../assets/data-img/item_9.jpg',
         rating: {
             rate: 3.3,
             count: 203,
