+++ src/pages/About/AboutPage.tsx
@@ -1,7 +1,7 @@
 import React, { Component } from 'react';
 import './AboutPage.css';
 
-export default class AboutPage extends Component {
+export class AboutPage extends Component {
     render() {
         return <div className="about">Lorem ipsum dolor sit amet consectetur adipisicing elit?</div>;
     }
