+++ src/pages/NotFound/NotFoundPage.tsx
@@ -1,7 +1,7 @@
 import React, { Component } from 'react';
 import './NotFoundPage.css';
 
-export default class NotFound extends Component {
+export class NotFoundPage extends Component {
     render() {
         return <h2 className="notfound">Oops! Not Found Page!</h2>;
     }
